ETDA-1518 Using new autoloader.  Fixed namespacing issue with eamil contact form, removed optional association with committee member toke, add optional association test with committee members, fixed bugged out tests.
