+++ spec/integration/admin/submissions/approve_format_review_spec.rb
@@ -1,0 +1,40 @@
+RSpec.describe "when an admin approves a format review", js: true do
+  require 'integration/integration_spec_helper'
+
+  let!(:author) { FactoryBot.create :author }
+  let!(:submission) { FactoryBot.create :submission, :waiting_for_format_review_response, author: author }
+
+  before do
+    webaccess_authorize_admin
+  end
+
+  context "when an admin accepts the format review files" do
+    it "changes status to 'collecting final submission files'" do
+      expect(submission.format_review_approved_at).to be_nil
+      FactoryBot.create :format_review_file, submission: submission
+      visit admin_edit_submission_path(submission)
+      fill_in 'Format Review Notes to Student', with: 'Note on format review'
+      click_button 'Format Review Completed'
+      # expect(page).to have_content('successfully')
+      submission.reload
+      expect(submission.status).to eq 'collecting final submission files'
+      submission.reload
+      expect(submission.format_review_approved_at).not_to be_nil
+    end
+  end
+
+  context "when an admin rejects the format review files" do
+    it "changes status to 'collecting format review files rejected'" do
+      expect(submission.format_review_rejected_at).to be_nil
+      FactoryBot.create :format_review_file, submission: submission
+      visit admin_edit_submission_path(submission)
+      fill_in 'Format Review Notes to Student', with: 'Note on need for revisions'
+      click_button 'Reject & request revisions'
+      # expect(page).to have_content('successfully')
+      submission.reload
+      expect(submission.status).to eq 'collecting format review files rejected'
+      submission.reload
+      expect(submission.format_review_rejected_at).not_to be_nil
+    end
+  end
+end
